refactor: 1차 피드백 반영
