+++ backend/web/src/frontend/src/api/index.js
@@ -2,7 +2,7 @@ import axios from 'axios';
 
 const ApiService = {
   get(uri) {
-    return axios.get(`http://localhost:8080${uri}`);
+    return axios.get(uri);
   },
 };
