+++ backend/web/src/frontend/src/api/module/bookmark.js
@@ -1,4 +1,4 @@
-import ApiService from '../index.js';
+import ApiService from '@/api/index.js';
 
 const BASE_URL = '/api/v1/tags';
