+++ backend/web/src/frontend/src/store/share/getterTypes.js
@@ -1,0 +1,2 @@
+export const BOOKMARKS = 'bookmarks';
+export const CATEGORIES = 'categories';
