+++ backend/web/src/frontend/src/views/bookmark/BookmarkPage.vue
@@ -5,13 +5,13 @@
 <script>
 import BookmarkCard from '@/views/bookmark/components/BookmarkCard.vue';
 import { mapActions } from 'vuex';
-import { FETCH_BOOKMARKS } from '@/store/share/actionType.js';
+import { FETCH_BOOKMARKS } from '@/store/share/actionTypes.js';
 
 export default {
   name: 'BookmarkPage',
   beforeRouteUpdate(to, from, next) {
     next();
-    this.fetchBookmarks({ tagId: this.$route.params.id });
+    this[FETCH_BOOKMARKS]({ tagId: this.$route.params.id });
   },
   methods: {
     ...mapActions([FETCH_BOOKMARKS]),
