+++ backend/web/src/frontend/src/views/bookmark/components/BookmarkCard.vue
@@ -15,9 +15,10 @@
 
 <script>
 import { mapActions, mapGetters } from 'vuex';
-import { FETCH_BOOKMARKS } from '@/store/share/actionType.js';
-import CardStream from './CardStream.vue';
-import CardModule from './CardModule.vue';
+import { FETCH_BOOKMARKS } from '@/store/share/actionTypes.js';
+import { BOOKMARKS } from '@/store/share/getterTypes.js';
+import CardStream from '@/views/bookmark/components/CardStream.vue';
+import CardModule from '@/views/bookmark/components/CardModule.vue';
 
 export default {
   name: 'BookmarkCard',
@@ -26,7 +27,7 @@ export default {
     CardModule,
   },
   computed: {
-    ...mapGetters(['bookmarks']),
+    ...mapGetters([BOOKMARKS]),
   },
   data() {
     return {
