+++ backend/web/src/frontend/src/views/bookmark/components/CardIcons.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script>
-import TagEditModal from './TagEditModal.vue';
+import TagEditModal from '@/views/bookmark/components/TagEditModal.vue';
 
 export default {
   name: 'CardIcons',
@@ -27,5 +27,3 @@ export default {
   },
 };
 </script>
-
-<style scoped></style>
