+++ backend/web/src/frontend/src/views/bookmark/components/TagEditModal.vue
@@ -54,5 +54,3 @@ export default {
   },
 };
 </script>
-
-<style scoped></style>
